+++ course/src/main/java/in/testpress/course/util/PinchToZoomGesture.kt
@@ -1,20 +1,17 @@
 package `in`.testpress.course.util
 
 import `in`.testpress.course.R
+import `in`.testpress.course.util.extension.animateHideTextView
 import android.content.Context
 import android.os.*
-import android.view.MotionEvent
+import android.util.Log
 import android.view.ScaleGestureDetector
 import android.view.ScaleGestureDetector.SimpleOnScaleGestureListener
 import android.view.View
-import android.view.View.OnTouchListener
-import android.view.animation.*
 import android.widget.TextView
-import androidx.core.view.isVisible
 import com.github.vkay94.dtpv.DoubleTapPlayerView
 import com.google.android.exoplayer2.ui.AspectRatioFrameLayout
 import java.text.DecimalFormat
-
 
 class PinchToZoomGesture(
     exoPlayerUtil: ExoPlayerUtil,
@@ -30,6 +27,9 @@ class PinchToZoomGesture(
 
     override fun onScale(detector: ScaleGestureDetector): Boolean {
         playerView.hideController()
+        Log.d("TAG", "onScale: ${detector.scaleFactor}")
+        // Update the scaleFactor by multiplying it with the detected scale factor,
+        // and ensure it stays within the range of 1.0f to 6.0f.
         scaleFactor *= detector.scaleFactor
         scaleFactor = 1.0f.coerceAtLeast(scaleFactor.coerceAtMost(6.0f))
         playerView.videoSurfaceView?.scaleX = scaleFactor
@@ -76,48 +76,13 @@ class PinchToZoomGesture(
     private fun updateZoomModeTextView() {
         zoomModeText.visibility = View.VISIBLE
         zoomModeText.text = currentMode.mode
-        zoomModeText.hideTextView()
+        zoomModeText.animateHideTextView()
     }
 
     private fun updateZoomSizeTextView() {
         zoomModeText.clearAnimation()
         zoomSizeText.visibility = View.VISIBLE
         zoomSizeText.text = DecimalFormat("0.0x").format(scaleFactor.toDouble())
-    }
-
-    private fun TextView.hideTextView() {
-        val animationSet = AnimationSet(false)
-
-        // Fade in animation
-        val fadeInAnimation = AlphaAnimation(0.0f, 1.0f)
-        fadeInAnimation.duration = 250
-        animationSet.addAnimation(AlphaAnimation(0.0f, 1.0f).apply {
-            duration = 250
-        })
-
-        // Stay visible for 500ms
-        val stayDuration = 500
-        animationSet.addAnimation(AlphaAnimation(1.0f, 1.0f).apply {
-            duration = stayDuration.toLong()
-            startOffset = fadeInAnimation.duration
-        })
-
-        // Fade out animation
-        val fadeOutAnimation = AlphaAnimation(1.0f, 0.0f)
-        fadeOutAnimation.duration = 250
-        fadeOutAnimation.startOffset = fadeInAnimation.duration + stayDuration
-        animationSet.addAnimation(fadeOutAnimation)
-
-        animationSet.setAnimationListener(object : Animation.AnimationListener {
-            override fun onAnimationStart(animation: Animation?) {}
-            override fun onAnimationEnd(animation: Animation?) {
-                this@hideTextView.visibility = View.GONE
-            }
-
-            override fun onAnimationRepeat(animation: Animation?) {}
-        })
-
-        this.startAnimation(animationSet)
     }
 
     fun resetPinchToZoomGesture(zoomMode: ZoomMode) {
@@ -139,66 +104,4 @@ class PinchToZoomGesture(
         ZOOM("Zoom")
     }
 
-}
-
-class OnTouchDragListener(private val exoPlayerUtil: ExoPlayerUtil) : OnTouchListener {
-
-    private var lastTouchX = 0f
-    private var lastTouchY = 0f
-    private var posX = 0f
-    private var posY = 0f
-    private var moveCalled = 0
-    private val playerView get() = exoPlayerUtil.playerView
-    private val scaleGesture get() = exoPlayerUtil.scaleGesture
-
-    override fun onTouch(p0: View, motionEvent: MotionEvent): Boolean {
-        if (exoPlayerUtil.fullscreen) {
-            exoPlayerUtil.scaleGestureDetector.onTouchEvent(motionEvent)
-            if (scaleGesture.isDragEnabled) {
-                when (motionEvent.action) {
-                    MotionEvent.ACTION_DOWN -> {
-                        lastTouchX = motionEvent.x
-                        lastTouchY = motionEvent.y
-                        return false
-                    }
-                    MotionEvent.ACTION_MOVE -> {
-                        moveCalled += 1
-                        if (moveCalled < 5) {
-                            return false
-                        }
-                        val deltaX: Float = motionEvent.rawX - lastTouchX
-                        val deltaY: Float = motionEvent.rawY - lastTouchY
-
-                        // Calculate the new positions
-                        val newPosX: Float = posX + deltaX
-                        val newPosY: Float = posY + deltaY
-
-                        // Calculate the maximum allowed translations
-                        val maxPosX: Float =
-                            (playerView.videoSurfaceView?.width!! * scaleGesture.scaleFactor - playerView.width) / 2
-                        val maxPosY: Float =
-                            (playerView.videoSurfaceView?.height!! * scaleGesture.scaleFactor - playerView.height) / 2
-                        val minPosX = -maxPosX
-                        val minPosY = -maxPosY
-
-                        // Apply boundary checks
-                        posX = newPosX.coerceAtLeast(minPosX).coerceAtMost(maxPosX)
-                        posY = newPosY.coerceAtLeast(minPosY).coerceAtMost(maxPosY)
-
-                        // Apply translations
-                        playerView.videoSurfaceView?.translationX = posX
-                        playerView.videoSurfaceView?.translationY = posY
-                        lastTouchX = motionEvent.rawX
-                        lastTouchY = motionEvent.rawY
-                        return true
-                    }
-                    MotionEvent.ACTION_UP -> {
-                        moveCalled = 0
-                        return false
-                    }
-                }
-            }
-        }
-        return false
-    }
 }
